Reversing Letters (Guess it's my last code).
What this code basically does is it takes the user's input and reverses the letters inside of it. The order still stays the same. Also, I'd like to thank everyone in Gloucester H.S. for supporting me in this short time which I enjoyed very much. Thanks for everything, Gators. Engineer out.